+++ tests/testthat/test-category.R
@@ -135,3 +135,11 @@ test_that("MCScorrect argument bounds the results to -1.8C", {
   expect_equal(as.numeric(table(cat$climatology$category)[1]), 479)
   expect_equal(as.numeric(table(cat_correct$climatology$category)[4]), 271)
 })
+
+test_that("MCSice creates a 'V Ice' category", {
+  sst_ice <- sst_NW_Atl; sst_ice$temp <- sst_ice$temp-5; sst_ice$temp <- ifelse(sst_ice$temp < -1.8, -1.8, sst_ice$temp)
+  ts_ice <- ts2clm(sst_ice, climatologyPeriod = c("1982-01-01", "2011-12-31"), pctile = 10)
+  res <- detect_event(ts_ice, coldSpells = T)
+  cat_ice <- category(res, climatology = T, MCScorrect = T, MCSice = T)
+  expect_equal(as.numeric(table(cat_ice$event$category)[5]), 13)
+})