+++ NEWS.md
@@ -1,6 +1,8 @@
 # heatwaveR 0.4.6.9001 (2023-01-17)
 
 -   Updated citations page for all references to `heatwaveR` from 2022
+-   Added package names to all functions used in the downloading OISST vignette
+-   Added the `MCSice` argument to `category()`, which allows users to directly classify a MCS with a threshold below -1.7°C as a category 'V Ice' event _sensu_ Schlegel et al. (2021; Marine cold-spells)
 
 # heatwaveR 0.4.6.9001 (2022-06-23)